Sparsity and Pruning schedulers implementation for torch was moved to common part
