+++ tests/pruning/test_common.py
@@ -12,7 +12,7 @@
 """
 import pytest
 
-from nncf.pruning.schedulers import BaselinePruningScheduler, ExponentialPruningScheduler, \
+from nncf.common.pruning.schedulers import BaselinePruningScheduler, ExponentialPruningScheduler, \
     ExponentialWithBiasPruningScheduler
 from tests.pruning.helpers import PruningTestModel, get_basic_pruning_config
 from tests.helpers import create_compressed_model_and_algo_for_test
