+++ tests/pruning/test_schedulers.py
@@ -13,7 +13,7 @@
 import numpy as np
 import pytest
 
-from nncf.pruning.schedulers import BaselinePruningScheduler, ExponentialWithBiasPruningScheduler
+from nncf.common.pruning.schedulers import BaselinePruningScheduler, ExponentialWithBiasPruningScheduler
 from tests.pruning.helpers import get_pruning_baseline_config, PruningTestModel, get_pruning_exponential_config
 from tests.helpers import create_compressed_model_and_algo_for_test
